Create first.py
In this file you can find a way to create an simple mininet_topo with python3.x, it include some modules of CPULimitedHost and TCLink. The CPUlimitedHost is used to liminet host's CPU which you had created. TCLink is used to limit the links' bandwidth, delay and max_queue_size...